src/requirements.txt
python-dotenv==1.0.1
telebot==0.0.5
rich==13.9.3
aiohttp==3.13.1

langchain==0.3.7
langchain-core==0.3.15
langchain-community==0.3.3
langgraph==0.1.18
langsmith==0.1.125

langchain-mcp-adapters==0.0.6
mcp==1.19.0
